Fix tests
We need the zuul routing section inside the test.properties. I have no
idea why this is mandatory because the test uses a different property
file (test-routing.properties). But spring seems to autowire the wrong
configuration inside the zuul filter. This only happens when running the
tests with maven and not with eclipse.